requirements.txt
streamlit
google-generativeai
pandas
pillow
pyngrok
